refactor(gateway): worker initialization moves out of the gateway root into transports/startup.py
