+++ gateway/startup.py
@@ -1,122 +1,34 @@
 """Start and stop everything the gateway serves users through.
 
-The one file allowed to import every transport: it holds the transport
-registry and composes web + chat into a single running handle. Mirrors each
-transport's own ``startup.py`` one level up — those start one platform, this
-starts the gateway.
+The facade the process controller calls: one :func:`start_gateway` that brings
+up the web server and every chat transport together and returns the running
+handle. Worker initialization lives with the transports
+(:mod:`gateway.transports.startup`); this module only composes.
 
-Only the composition root (:class:`~gateway.core.runtime.controller.GatewayController`)
-imports this module. Transports never import it back, and ``gateway.core``
-cannot (core must not depend on transports). The scheduler is not started
-here; the manager starts it as a peer.
+Only :class:`~gateway.core.runtime.controller.GatewayController` imports this
+module, and only this module imports ``gateway.transports.startup``.
 """
 
 from __future__ import annotations
 
 import logging
-from collections.abc import Sequence
 from dataclasses import dataclass, field
 
-from gateway.core.runtime.errors import (
-    GatewayConfigurationError,
-    GatewayTransportFailedError,
+from gateway.core.transport_api import GatewayAgentCallback, TransportName
+from gateway.transports.startup import (
+    DEFAULT_STOP_TIMEOUT_SECONDS,
+    TransportHandle,
+    start_transports,
+    stop_transports,
 )
-from gateway.core.transport_api import (
-    GatewayAgentCallback,
-    TransportName,
-    TransportSpec,
-    TransportWorker,
-)
-from gateway.transports.buzz.startup import start_buzz_worker
-from gateway.transports.discord.startup import start_discord_worker
-from gateway.transports.slack.startup import start_slack_worker
-from gateway.transports.telegram.startup import start_telegram_worker
 from gateway.web.startup import start_web_server
 from gateway.web.web_server import WebAppServerHandle
-
-# How long a shutdown waits on each worker before giving up on it.
-DEFAULT_STOP_TIMEOUT_SECONDS = 8.0
 
 # The web app is a thread join, not a network drain, so it gets a smaller slice
 # of the shutdown budget and leaves the rest for in-flight chat turns.
 WEB_STOP_TIMEOUT_SECONDS = 5.0
 
 _WEB_COMPONENT = "web"
-
-
-TRANSPORTS: tuple[TransportSpec, ...] = (
-    TransportSpec(TransportName.TELEGRAM, start_telegram_worker, "polling for messages"),
-    TransportSpec(TransportName.SLACK, start_slack_worker, "inbound connected"),
-    TransportSpec(TransportName.DISCORD, start_discord_worker, "connected via gateway"),
-    TransportSpec(TransportName.BUZZ, start_buzz_worker, "polling for messages"),
-)
-
-
-@dataclass(frozen=True)
-class TransportHandle:
-    """A started transport: the worker to stop, and how to describe it."""
-
-    name: TransportName
-    worker: TransportWorker
-    status: str
-
-
-@dataclass(frozen=True)
-class ChatStartup:
-    """Started chat transports plus the status of every transport that was tried.
-
-    ``statuses`` covers transports that did not start, so the caller can report
-    "not configured" without the callee reaching into its status map.
-    """
-
-    handles: list[TransportHandle]
-    statuses: dict[TransportName, str]
-
-
-def start_transports(
-    *,
-    logger: logging.Logger,
-    handler: GatewayAgentCallback,
-) -> ChatStartup:
-    """Start every configured transport and report what each one did.
-
-    * :class:`GatewayConfigurationError` → ``not configured (…)`` (skipped).
-    * :class:`GatewayTransportFailedError` → ``failed (…)`` (skipped).
-
-    The gateway still serves whichever transports started successfully.
-    """
-    handles: list[TransportHandle] = []
-    statuses: dict[TransportName, str] = {}
-    for spec in TRANSPORTS:
-        try:
-            worker, _settings = spec.start(logger=logger, handler=handler)
-        except GatewayConfigurationError as exc:
-            logger.warning("%s chat disabled: %s", spec.name.capitalize(), exc)
-            statuses[spec.name] = f"not configured ({exc})"
-            continue
-        except GatewayTransportFailedError as exc:
-            logger.warning("%s chat failed: %s", spec.name.capitalize(), exc)
-            statuses[spec.name] = f"failed ({exc})"
-            continue
-        handles.append(TransportHandle(name=spec.name, worker=worker, status=spec.running_status))
-        statuses[spec.name] = spec.running_status
-    return ChatStartup(handles=handles, statuses=statuses)
-
-
-def stop_transports(
-    *,
-    handles: Sequence[TransportHandle],
-    timeout: float = DEFAULT_STOP_TIMEOUT_SECONDS,
-) -> bool:
-    """Stop every started transport and return whether all of them stopped.
-
-    Every worker is asked to stop even after one fails, so a single stuck
-    transport cannot leave the others running.
-    """
-    stopped = True
-    for handle in handles:
-        stopped = handle.worker.stop(timeout=timeout) and stopped
-    return stopped
 
 
 @dataclass
@@ -161,12 +73,7 @@ def start_gateway(
 
 __all__ = [
     "DEFAULT_STOP_TIMEOUT_SECONDS",
-    "TRANSPORTS",
     "WEB_STOP_TIMEOUT_SECONDS",
-    "ChatStartup",
     "StartedGateway",
-    "TransportHandle",
     "start_gateway",
-    "start_transports",
-    "stop_transports",
 ]
